requirements.txt
openai>=1.30.0
streamlit>=1.35.0
chromadb>=0.5.0
pypdf2>=3.0.0
pydantic>=2.0.0
langgraph>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
celery>=5.3.0
redis>=5.0.0
tenacity>=8.2.0
